requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
compile==1.0.3
configparser==5.0.2
crayons==0.4.0
cryptography==3.4.5
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.10
django-allauth==0.44.0
django-circles==0.1.0
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-email-manager==0.2.1
django-heroku==0.3.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-post-office==3.5.3
django-posts==0.3.0
django-shapes==0.2.0
django-tinymce==3.2.0
et-xmlfile==1.0.1
eth-hash==0.3.1
eth-typing==2.2.2
figlet==0.0.1
filelock==3.0.12
gunicorn==20.1.0
humanfriendly==9.1
idna==2.10
jsonfield==3.1.0
MarkupPy==1.14
mysql-connector-python==8.0.16
mysqlclient==2.0.3
ndg-httpsclient==0.5.1
numpy==1.20.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
passlib==1.7.4
Pillow==8.1.0
protobuf==3.7.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pyfiglet==0.8.post1
PyJWT==2.0.1
pyOpenSSL==20.0.1
pyreadline==2.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2019.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.12.0
soupsieve==2.2
sqlparse==0.3.0
tablib==3.0.0
toml==0.10.2
toolz==0.11.1
Unipath==1.1
urllib3==1.26.3
virtualenv==20.4.2
webdriver-manager==3.3.0
whitenoise==5.2.0
wxPython==4.1.1
xlrd==2.0.1
xlwt==1.3.0
